Turn on CF_TRACE for these deployments;
Since this is an automated test that we generally only pay attention to
when there is something wrong with the platform, we should have CF_TRACE
set to true all the time in case things break. Sure it's noisy, but that
noise is helpful when we're debugging and troubleshooting issues
customer's may be experiencing.